Dep-eject sheet.display
