+++ lib/depject/channel/html/preview.js
@@ -9,7 +9,6 @@ exports.needs = nest({
   'about.obs.name': 'first',
   'about.html.image': 'first',
   'keys.sync.id': 'first',
-  'sheet.display': 'first',
   'app.navigate': 'first',
   'intl.sync.i18n': 'first',
   'intl.sync.i18n_n': 'first',
