+++ lib/depject/gathering/sheet/edit.js
@@ -2,13 +2,13 @@ const nest = require('depnest')
 const extend = require('xtend')
 const Pickr = require('flatpickr')
 const spacetime = require('spacetime')
+const displaySheet = require('../../../sheet/display')
 
 const { Value, h, computed, when } = require('mutant')
 
 exports.gives = nest('gathering.sheet.edit')
 
 exports.needs = nest({
-  'sheet.display': 'first',
   'message.sheet.preview': 'first',
   'keys.sync.id': 'first',
   'sbot.async.publish': 'first',
@@ -37,7 +37,7 @@ exports.create = function (api) {
       // use private gathering recps for default profile suggestions
       const participants = msg.value && msg.value.content && msg.value.content.recps
 
-      api.sheet.display(close => {
+      displaySheet(close => {
         const publishing = Value(false)
 
         const chosen = {
