+++ lib/depject/index.js
@@ -181,7 +181,6 @@ module.exports = {
     },
     sbot: require('./sbot.js'),
     sheet: {
-      display: require('./sheet/display.js'),
       editTags: require('./sheet/editTags.js'),
       profiles: require('./sheet/profiles.js'),
       tags: {
