+++ lib/depject/invite/sheet.js
@@ -1,9 +1,9 @@
 const { h, when, Value, Proxy } = require('mutant')
 const nest = require('depnest')
 const electron = require('electron')
+const displaySheet = require('../../sheet/display')
 
 exports.needs = nest({
-  'sheet.display': 'first',
   'invite.async.accept': 'first',
   'intl.sync.i18n': 'first'
 })
@@ -13,7 +13,7 @@ exports.gives = nest('invite.sheet')
 exports.create = function (api) {
   const i18n = api.intl.sync.i18n
   return nest('invite.sheet', function () {
-    api.sheet.display(close => {
+    displaySheet(close => {
       const publishing = Value()
       const publishStatus = Proxy()
