+++ lib/depject/message/sheet/preview.js
@@ -2,9 +2,9 @@ const h = require('mutant/h')
 const nest = require('depnest')
 const when = require('mutant/when')
 const emoji = require('node-emoji')
+const displaySheet = require('../../../sheet/display')
 
 exports.needs = nest({
-  'sheet.display': 'first',
   'message.html.render': 'first',
   'intl.sync.i18n': 'first',
   'intl.sync.i18n_n': 'first'
@@ -17,7 +17,7 @@ exports.create = function (api) {
   const plural = api.intl.sync.i18n_n
 
   return nest('message.sheet.preview', function (msg, cb) {
-    api.sheet.display(function (close) {
+    displaySheet(function (close) {
       const isPrivate = msg.value.private
       const isRoot = !msg.value.content.root
       const exists = !!msg.key
