+++ lib/depject/page/html/render/profile.js
@@ -1,6 +1,7 @@
 const nest = require('depnest')
 const ref = require('ssb-ref')
 const { h, when, computed, map, send, dictToCollection, resolve, onceTrue, Value } = require('mutant')
+const displaySheet = require('../../../../sheet/display')
 
 exports.needs = nest({
   'about.obs': {
@@ -24,7 +25,6 @@ exports.needs = nest({
   'sbot.async.getLatest': 'first',
   'sbot.obs.connection': 'first',
   'keys.sync.id': 'first',
-  'sheet.display': 'first',
   'profile.sheet.edit': 'first',
   'app.navigate': 'first',
   'profile.obs.contact': 'first',
@@ -381,7 +381,7 @@ exports.create = function (api) {
   }
 
   function rename (id) {
-    api.sheet.display(close => {
+    displaySheet(close => {
       const currentName = api.about.obs.name(id)
       const input = h('input', {
         style: { 'font-size': '150%' },
