+++ lib/depject/profile/html/preview.js
@@ -11,7 +11,6 @@ exports.needs = nest({
   'about.obs.description': 'first',
   'about.html.image': 'first',
   'keys.sync.id': 'first',
-  'sheet.display': 'first',
   'app.navigate': 'first',
   'intl.sync.i18n': 'first',
   'intl.sync.i18n_n': 'first',
