+++ lib/depject/profile/sheet/edit.js
@@ -3,11 +3,11 @@ const nest = require('depnest')
 const extend = require('xtend')
 const { Value, h, computed, when } = require('mutant')
 const fallbackImageUrl = 'data:image/gif;base64,R0lGODlhAQABAPAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw=='
+const displaySheet = require('../../../sheet/display')
 
 exports.gives = nest('profile.sheet.edit')
 
 exports.needs = nest({
-  'sheet.display': 'first',
   'keys.sync.id': 'first',
   'message.async.publish': 'first',
   'sbot.async.publish': 'first',
@@ -20,14 +20,14 @@ exports.needs = nest({
   'blob.html.input': 'first',
   'blob.sync.url': 'first',
   'intl.sync.i18n': 'first',
-  'suggest.hook': 'first',
+  'suggest.hook': 'first'
 })
 
 exports.create = function (api) {
   const i18n = api.intl.sync.i18n
   return nest('profile.sheet.edit', function ({ usePreview = true } = {}) {
     const id = api.keys.sync.id()
-    api.sheet.display(close => {
+    displaySheet(close => {
       const currentName = api.about.obs.name(id)
       const currentImage = api.about.obs.image(id)
       const currentDescription = api.about.obs.description(id)
