+++ lib/depject/sheet/editTags.js
@@ -7,6 +7,7 @@ const addSuggest = require('suggest-box')
 const TagHelper = require('scuttle-tag')
 const ref = require('ssb-ref')
 const catchLinks = require('../../catch-links')
+const displaySheet = require('../../sheet/display')
 
 exports.gives = nest('sheet.editTags')
 
@@ -14,7 +15,6 @@ exports.needs = nest({
   'about.obs.name': 'first',
   'keys.sync.id': 'first',
   'sbot.obs.connection': 'first',
-  'sheet.display': 'first',
   'tag.async.suggest': 'first',
   'tag.html.tag': 'first',
   'intl.sync.i18n': 'first',
@@ -32,7 +32,7 @@ exports.create = function (api) {
 
     callback = callback || function () {}
 
-    api.sheet.display(function (close) {
+    displaySheet(function (close) {
       const { content, onMount, onSave } = edit({ msgId })
 
       const wrapper = h('div', [
