+++ lib/depject/sheet/profiles.js
@@ -1,9 +1,9 @@
 const { h, map, computed, Value, lookup } = require('mutant')
 const nest = require('depnest')
 const catchLinks = require('../../catch-links')
+const displaySheet = require('../../sheet/display')
 
 exports.needs = nest({
-  'sheet.display': 'first',
   'keys.sync.id': 'first',
   'contact.obs.following': 'first',
   'contact.html.followToggle': 'first',
@@ -19,7 +19,7 @@ exports.gives = nest('sheet.profiles')
 exports.create = function (api) {
   const i18n = api.intl.sync.i18n
   return nest('sheet.profiles', function (ids, title) {
-    api.sheet.display(close => {
+    displaySheet(close => {
       const currentFilter = Value()
       const nameLookup = lookup(ids, (id) => {
         return [id, api.about.obs.name(id)]
