+++ lib/depject/sheet/tags/tags.js
@@ -1,11 +1,11 @@
 const { h, computed, Value } = require('mutant')
 const nest = require('depnest')
 const catchLinks = require('../../../catch-links')
+const displaySheet = require('../../../sheet/display')
 
 exports.needs = nest({
   'app.navigate': 'first',
   'intl.sync.i18n': 'first',
-  'sheet.display': 'first',
   'sheet.tags.renderTags': 'first',
   'sheet.tags.renderTaggers': 'first'
 })
@@ -17,7 +17,7 @@ exports.create = function (api) {
   const displayTags = Value(true)
   const selectedTag = Value()
   return nest('sheet.tags.render', function (msgId, ids) {
-    api.sheet.display(close => {
+    displaySheet(close => {
       const content = h('div',
         computed([displayTags, selectedTag], (display, tag) => {
           if (!display) return api.sheet.tags.renderTaggers(msgId, tag)
