+++ locales/en.json
@@ -303,5 +303,6 @@
 	"Connections": "Connections",
 	"Connect": "Connect",
 	"In order to share with users on the internet, you need to be invited to a pub server or a room server.": "In order to share with users on the internet, you need to be invited to a pub server or a room server.",
-	"nl": "nl"
+	"nl": "nl",
+	"Last activity": "Last activity"
 }
